+++ src/Tools/AutoAccept.cs
@@ -1,7 +1,7 @@
 ﻿using Loly.src.Logs;
 using Loly.src.Variables;
 using Loly.src.Variables.Enums;
-using static Loly.src.LeagueClient.Requests;
+using static Loly.src.Tools.Requests;
 
 namespace Loly.src.Tools;
